+++ internal/provider/resource_tfe_policy_set_parameter_test.go
@@ -12,8 +12,11 @@ import (
 
 	tfe "github.com/hashicorp/go-tfe"
 	"github.com/hashicorp/go-version"
+	"github.com/hashicorp/terraform-plugin-testing/compare"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/statecheck"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
+	"github.com/hashicorp/terraform-plugin-testing/tfjsonpath"
 	"github.com/hashicorp/terraform-plugin-testing/tfversion"
 )
 
@@ -148,6 +151,9 @@ func TestAccTFEPolicySetParameter_valueWO(t *testing.T) {
 	paramValue1 := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 	paramValue2 := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
+	// Create the value comparer so we can add state values to it during the test steps
+	compareValuesDiffer := statecheck.CompareValue(compare.ValuesDiffer())
+
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() { testAccPreCheck(t) },
 		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
@@ -156,11 +162,11 @@ func TestAccTFEPolicySetParameter_valueWO(t *testing.T) {
 		ProtoV5ProviderFactories: testAccMuxedProviders,
 		CheckDestroy:             testAccCheckTFEPolicySetParameterDestroy,
 		Steps: []resource.TestStep{
-			{
+			{ // Should not be able to set both `value` and `value_wo` simultaneously
 				Config:      testAccTFEPolicySetParameter_valueAndValueWO(org.Name, paramValue1),
 				ExpectError: regexp.MustCompile(`Attribute "value" cannot be specified when "value_wo" is specified`),
 			},
-			{
+			{ // Provision a sensitive parameter with a write-only value
 				Config: testAccTFEPolicySetParameter_valueWO(org.Name, paramValue1, true),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckTFEPolicySetParameterExists(
@@ -170,8 +176,15 @@ func TestAccTFEPolicySetParameter_valueWO(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"tfe_policy_set_parameter.foobar", "sensitive", "true"),
 				),
+				// Register the id with the value comparer so we can assert that the
+				// resource has been replaced in the next step.
+				ConfigStateChecks: []statecheck.StateCheck{
+					compareValuesDiffer.AddStateValue(
+						"tfe_policy_set_parameter.foobar", tfjsonpath.New("id"),
+					),
+				},
 			},
-			{
+			{ // Update the value of the write-only parameter and set sensitive: false
 				Config: testAccTFEPolicySetParameter_valueWO(org.Name, paramValue2, false),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckTFEPolicySetParameterExists(
@@ -181,6 +194,12 @@ func TestAccTFEPolicySetParameter_valueWO(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"tfe_policy_set_parameter.foobar", "sensitive", "false"),
 				),
+				// Ensure that the resource has been replaced
+				ConfigStateChecks: []statecheck.StateCheck{
+					compareValuesDiffer.AddStateValue(
+						"tfe_policy_set_parameter.foobar", tfjsonpath.New("id"),
+					),
+				},
 			},
 			{
 				Config: testAccTFEPolicySetParameter_basic(org.Name),
@@ -194,6 +213,12 @@ func TestAccTFEPolicySetParameter_valueWO(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"tfe_policy_set_parameter.foobar", "sensitive", "false"),
 				),
+				// Ensure that the resource has been replaced
+				ConfigStateChecks: []statecheck.StateCheck{
+					compareValuesDiffer.AddStateValue(
+						"tfe_policy_set_parameter.foobar", tfjsonpath.New("id"),
+					),
+				},
 			},
 		},
 	})